requirements.txt
albumentations==1.1.0
numpy==1.22.3
pandas==1.4.2
Pillow==9.4.0
PyYAML==6.0
scikit_learn==1.2.1
scipy==1.8.1
torch==1.11.0
tqdm==4.64.0
wandb==0.12.17
